refactor: split model into entity and domain
