+++ src/main/kotlin/omok/model/domain/rule/ThreeThreeRule.kt
@@ -1,4 +1,4 @@
-package omok.model.rule
+package omok.model.domain.rule
 
 object ThreeThreeRule : Rule() {
     override fun validated(
