+++ src/main/kotlin/omok/model/domain/rule/WhiteWinRule.kt
@@ -1,12 +1,12 @@
-package omok.model.rule
+package omok.model.domain.rule
 
-object BlackWinRule : Rule() {
+object WhiteWinRule : Rule(WHITE_STONE, BLACK_STONE) {
     override fun validated(
         board: List<List<Int>>,
         position: Pair<Int, Int>,
-    ): Boolean = directions.map { direction -> checkBlackWin(board, position, direction) }.contains(true)
+    ): Boolean = directions.map { direction -> checkWhiteWin(board, position, direction) }.contains(true)
 
-    private fun checkBlackWin(
+    private fun checkWhiteWin(
         board: List<List<Int>>,
         position: Pair<Int, Int>,
         direction: Pair<Int, Int>,
@@ -16,7 +16,7 @@ object BlackWinRule : Rule() {
         val (stone2, blink2) = search(board, position, direction)
 
         return when {
-            blink1 + blink2 == 0 && stone1 + stone2 == 4 -> true
+            blink1 + blink2 == 0 && stone1 + stone2 >= 4 -> true
             else -> false
         }
     }
