+++ src/main/kotlin/omok/model/entity/Stone.kt
@@ -1,4 +1,4 @@
-package omok.model
+package omok.model.entity
 
 enum class Stone {
     BLACK,
