+++ src/test/kotlin/omok/model/BoardPositionStateTest.kt
@@ -1,6 +1,7 @@
 package omok.model
 
-import omok.model.board.BoardPositionState
+import omok.model.entity.Stone
+import omok.model.entity.board.BoardPositionState
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertAll
