+++ src/test/kotlin/omok/model/BoardPositionTest.kt
@@ -1,8 +1,8 @@
 package omok.model
 
-import omok.model.board.BoardPosition
-import omok.model.board.BoardPositionState
-import omok.model.board.DefaultBoardPosition
+import omok.model.entity.board.BoardPosition
+import omok.model.entity.board.BoardPositionState
+import omok.model.entity.board.DefaultBoardPosition
 import omok.model.testDouble.FakePosition
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
