+++ src/test/kotlin/omok/model/StoneTest.kt
@@ -1,5 +1,6 @@
 package omok.model
 
+import omok.model.entity.Stone
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
