+++ src/test/kotlin/omok/model/omokRule/OmokRuleWhiteWinTest.kt
@@ -1,6 +1,6 @@
 package omok.domain.omokRule
 
-import omok.model.rule.WhiteWinRule
+import omok.model.domain.rule.WhiteWinRule
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
